requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.49
botocore==1.29.49
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
Django==4.1.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
validators==0.20.0
